demos/stqueue-pydemo/requirement.txt
azure-storage-queue 
azure-identity